Remove unnecessary params from API call.
The branch names are now fetched automatically on the backend.

The commit hash is also optional now, but it's a best practice to
include it when it's available because the images that CI has just built
might not correspond to the true head of the branch by the time it has
finished building them.